requirements.txt
# 동의보감 RAG 시스템 필수 라이브러리 (OpenAI 전용)
# pip install -r requirements.txt

# 핵심 AI/ML 라이브러리
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
numpy>=1.24.0
pandas>=2.0.0

# 중국어 처리
jieba>=0.42.1
opencc-python-reimplemented>=0.1.7

# LangChain 관련 (최신 버전)
langchain>=0.1.0
langchain-community>=0.0.25
langchain-core>=0.1.0
tiktoken>=0.5.0
openai>=1.0.0

# 환경 설정
python-dotenv>=0.19.0

# 기타 유틸리티
pathlib2>=2.3.7

# 개발 도구 (선택사항)
jupyter>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
